Moved chromatic tests to input folder & added a transparent prop to the custom Input component
