+++ packages/react-components/components/input/src/input.jsx
@@ -1,14 +1,31 @@
 import { Input as SemanticInput } from "semantic-ui-react";
+import { bool, oneOf, string } from "prop-types";
 import { isNil } from "lodash";
-import { oneOf } from "prop-types";
-import { throwWhenUnsupportedPropIsProvided } from "@orbit-ui/react-components-shared";
+import { mergeClasses, throwWhenUnsupportedPropIsProvided } from "@orbit-ui/react-components-shared";
 
 const UNSUPPORTED_PROPS = [];
 
-export function Input({ children, ...props }) {
+const propTypes = {
+    transparent: bool,
+    /**
+     * @ignore
+     */
+    className: string
+};
+
+const defaultProps = {
+    transparent: false
+};
+
+export function Input({ transparent, className, children, ...props }) {
     throwWhenUnsupportedPropIsProvided(props, UNSUPPORTED_PROPS);
 
-    return <SemanticInput {...props}>{children}</SemanticInput>;
+    const classes = mergeClasses(
+        transparent && "transparent",
+        className
+    );
+
+    return <SemanticInput className={classes} {...props}>{children}</SemanticInput>;
 }
 
 // eslint-disable-next-line react/forbid-foreign-prop-types
@@ -16,3 +33,6 @@ if (!isNil(SemanticInput.propTypes)) {
     // eslint-disable-next-line react/forbid-foreign-prop-types
     SemanticInput.propTypes.size = oneOf(["tiny", "small", "medium", "large"]);
 }
+
+Input.propTypes = propTypes;
+Input.defaultProps = defaultProps;
