+++ packages/react-components/components/input/stories/input.stories.mdx
@@ -1,6 +1,7 @@
+import { HocWarning, Installation, ReactComponentsLayout } from "@blocks/react-components";
 import { Input } from "@orbit-ui/react-input/src";
-import { Installation, ReactComponentsLayout } from "@blocks/react-components";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
+import { Props } from "@blocks";
 import { createComponentSection, paramsBuilder } from "@utils";
 import Imports from "./imports.mdx";
 
@@ -17,6 +18,16 @@ import Imports from "./imports.mdx";
 ## Installation
 
 <Installation componentFolder="input" imports={<Imports />} />
+
+## Props
+
+<HocWarning componentName="Input" semanticPath="/elements/input" />
+
+<Props
+    componentsDefinitions={[
+        { displayName: "Input", component: Input }
+    ]}
+/>
 
 ## Usage
 
@@ -155,7 +166,7 @@ An input can be transparent.
 
 <Preview>
     <Story name="transparent">
-        <Input className="transparent" placeholder="Search..." />
+        <Input transparent placeholder="Search..." />
     </Story>
 </Preview>
