+++ packages/react-components/components/search-input/tests/chromatic/search-input-controller.chroma.jsx
@@ -39,6 +39,12 @@ stories()
              createSearchInputController({
                  autofocus: true
              })
+    )
+    .add("fluid",
+         () =>
+             createSearchInputController({
+                 fluid: true
+             })
     );
 
 stories("/loading/closed")
