+++ storybook/stories/introduction/tachyons.stories.mdx
@@ -39,10 +39,6 @@ Orbit provide a custom version of the Tachyons library based on [Orbit foundatio
 
 Most of the documentation is available on [Tachyons website](http://tachyons.io/docs). For Orbit customization documentation, refer to the [usage section](#usage).
 
-## Who should use Tachyons?
-
-Every ShareGate's products should use Tachyons as a base for their CSS foundation. If you are using one of [Orbit React components](?path=/docs/introduction-packages--page#react-components) you are already using Tachyons.
-
 ## Installation
 
 Please refer to the [installation page](?path=/docs/introduction-installation--page) for instructions.
