--- storybook/stories/semantic-ui/react/input/imports.mdx
@@ -1,3 +1,0 @@
-```javascript
-    import { Input } from "semantic-ui-react";
-```
