--- storybook/stories/semantic-ui/react/input/input.stories.mdx
@@ -1,161 +1,0 @@
-import { Input } from "semantic-ui-react";
-import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
-import { SemanticLayout } from "../semantic-layout";
-import { createComponentSection, paramsBuilder } from "@utils";
-import Imports from "./imports.mdx";
-import Installation from "../installation.mdx";
-
-<Meta
-    title={createComponentSection("Input")}
-    parameters={paramsBuilder()
-        .docsLayout(<SemanticLayout documentationPath="/elements/input" />)
-        .build()}
-/>
-
-# Input
-
-## Installation
-
-<Installation imports={<Imports />} />
-
-## Usage
-
-### Default
-
-A default input.
-
-<Preview>
-    <Story name="default">
-        <Input />
-    </Story>
-</Preview>
-
-### Value
-
-An input can have an value.
-
-<Preview>
-    <Story name="value">
-        <Input value="Orbit" />
-    </Story>
-</Preview>
-
-### Focused
-
-An input can show a user is currently interacting with it.
-
-<Preview>
-    <Story name="focused">
-        <Input focus placeholder="Search..." />
-    </Story>
-</Preview>
-
-### Disabled
-
-An input can be disabled.
-
-<Preview>
-    <Story name="disabled">
-        <Input disabled placeholder="Search..." />
-    </Story>
-</Preview>
-
-### Fluid
-
-An input can take the width of its container.
-
-<Preview>
-    <Story name="fluid">
-        <Input fluid placeholder="Search..." />
-    </Story>
-</Preview>
-
-### Icon
-
-An input can be formatted with an icon.
-
-<Preview>
-    <Story name="icon">
-        <div className="flex flex-row">
-            <div className="mr2">
-                <Input icon="search" placeholder="Search..." />
-            </div>
-            <div>
-                <Input iconPosition="left" icon="search" placeholder="Search..." />
-            </div>
-        </div>
-    </Story>
-</Preview>
-
-### Loading
-
-An input can show that it is currently loading data.
-
-<Preview>
-    <Story name="loading">
-        <div className="flex flex-row">
-            <div className="mr2">
-                <Input loading placeholder="Search..." />
-            </div>
-            <div>
-                <Input loading iconPosition="left" placeholder="Search..." />
-            </div>
-        </div>
-    </Story>
-</Preview>
-
-### Size
-
-An input can vary in size.
-
-The available sizes are: `"tiny"`, `"small"`, `"medium"` and `"large"`.
-
-<Preview>
-    <Story name="size">
-        <div className="flex flex-row items-end">
-            <div className="flex-auto mr2">
-                <Input size="tiny" fluid placeholder="Search..." />
-            </div>
-            <div className="flex-auto mr2">
-                <Input size="small" fluid placeholder="Search..." />
-            </div>
-            <div className="flex-auto mr2">
-                <Input fluid placeholder="Search..." />
-            </div>
-            <div className="flex-auto mr2">
-                <Input size="large" fluid placeholder="Search..." />
-            </div>
-        </div>
-    </Story>
-</Preview>
-
-### Error
-
-An input field can show that the data contains errors.
-
-<Preview>
-    <Story name="error">
-        <Input error placeholder="Search..." />
-    </Story>
-</Preview>
-
-### Success
-
-An input field can show that the data has been validated.
-
-<Preview>
-    <Story name="success">
-        <Input className="success" placeholder="Search..." />
-    </Story>
-</Preview>
-
-### Transparent
-
-An input can be transparent.
-
-<Preview>
-    <Story name="transparent">
-        <Input className="transparent" placeholder="Search..." />
-    </Story>
-</Preview>
-
